+++ app/components/data-menu/layer-directives/eventseries-directive.js
@@ -34,6 +34,7 @@ angular.module('data-menu')
             uuid: scope.layer.uuid,
             url: 'api/v2/events/?event_series=' + scope.layer.uuid,
             aggType: response.aggregation_type,
+            color: response.color,
             scale: response.observation_type
               && response.observation_type.scale,
             quantity: response.observation_type